Battle service logic (#4)
* Added controllers, repositories and services

* Fixed some comments

* Removed authorization and room, added status PENDING

* Changed settings on embeddable

* Fixed comments

* Fixed comments

* Added part of battle service logic

* Implement methods in GameController and GameService

* Added check win logic

* Implement ToDo tasks

* Added missing UserRepository

* Implemented and added new ToDo comments

* Implemented ToDo comments

* Fixed some details by comments

* Removed link

* Changed the way of finding games

* Fixed some comments

* Added 2 services instead GameUtils and added SearchGameRequestDto, PreferableSide

* Added specifications

+++ common/src/main/java/com/company/common/dtos/UserDto.java
@@ -20,7 +20,7 @@ public class UserDto {
 
     @NotNull
     @Positive
-    private float rating;
+    private Integer rating;
 
     @NotBlank
     private String salt;
